hw/hw01/requirements.txt
otter-grader>=5.0.0
numpy>=1.21.0
matplotlib>=3.4.0
